Fix delete game wiping platform folder and drawer scroll
- Use canonicalPath for path comparison in DeleteGameUseCase
  Also require parent to be inside platform folder before deleting
  Fixes #83

- Make drawer menu items scrollable on tall/narrow screens
  Partial fix for #75

+++ app/src/main/kotlin/com/nendo/argosy/domain/usecase/game/DeleteGameUseCase.kt
@@ -62,10 +62,13 @@ class DeleteGameUseCase @Inject constructor(
                     file.deleteRecursively()
                 } else {
                     val parent = file.parentFile
-                    val isPlatformFolder = parent != null &&
-                        parent.absolutePath == platformFolder.absolutePath
+                    val platformCanonical = platformFolder.canonicalPath
+                    val parentCanonical = parent?.canonicalPath
 
-                    if (parent != null && !isPlatformFolder) {
+                    val isPlatformFolder = parentCanonical == platformCanonical
+                    val isInsidePlatformFolder = parentCanonical?.startsWith(platformCanonical) == true
+
+                    if (parent != null && !isPlatformFolder && isInsidePlatformFolder) {
                         parent.deleteRecursively()
                     } else {
                         file.delete()
